+++ collections/src/main/java/org/neo4j/gds/collections/HugeSparseDoubleArray.java
@@ -34,7 +34,7 @@ public interface HugeSparseDoubleArray {
         return growingBuilder(defaultValue, trackAllocation, 0);
     }
 
-    static HugeSparseDoubleArray.Builder growingBuilder(double defaultValue, LongConsumer trackAllocation, int initialCapacity) {
+    static HugeSparseDoubleArray.Builder growingBuilder(double defaultValue, LongConsumer trackAllocation, long initialCapacity) {
         return new HugeSparseDoubleArraySon.GrowingBuilder(defaultValue, trackAllocation, initialCapacity);
     }
