+++ collections/src/main/java/org/neo4j/gds/collections/HugeSparseIntArray.java
@@ -34,7 +34,7 @@ public interface HugeSparseIntArray {
         return growingBuilder(defaultValue, trackAllocation, 0);
     }
 
-    static Builder growingBuilder(int defaultValue, LongConsumer trackAllocation, int initialCapacity) {
+    static Builder growingBuilder(int defaultValue, LongConsumer trackAllocation, long initialCapacity) {
         return new HugeSparseIntArraySon.GrowingBuilder(defaultValue, trackAllocation, initialCapacity);
     }
