+++ collections/src/main/java/org/neo4j/gds/collections/HugeSparseShortArray.java
@@ -34,7 +34,7 @@ public interface HugeSparseShortArray {
         return growingBuilder(defaultValue, trackAllocation, 0);
     }
 
-    static Builder growingBuilder(short defaultValue, LongConsumer trackAllocation, int initialCapacity) {
+    static Builder growingBuilder(short defaultValue, LongConsumer trackAllocation, long initialCapacity) {
         return new HugeSparseShortArraySon.GrowingBuilder(defaultValue, trackAllocation, initialCapacity);
     }
